:recycle: BTS-300/refactor: TextEditor 슬라이스를 features에서 components 레벨로 이동 및 구조 응집
- 에디터 기능을 특정 도메인(feature)에 종속되지 않는 공용 복합 컴포넌트(슬라이스)로 재정의
- 위치 변경: features/editor -> components/editor
- 구조 개선: 내부 UI, 로직, 타입(model/types)을 세그먼트별로 분리
