+++ src/components/editor/index.ts
@@ -1,0 +1,5 @@
+export { TextEditor } from '@/components/editor/ui/text-editor';
+export { TextViewer } from '@/components/editor/ui/text-viewer';
+export { initialTextEditorValue } from '@/components/editor/model/initial-state';
+export { useTextEditor } from '@/components/editor/model/use-editor';
+export type * from './model/types';
