+++ src/components/editor/model/use-editor.ts
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 
-import { TextEditorValue, initialTextEditorValue } from '../utils';
+import { initialTextEditorValue } from '@/components/editor/model/initial-state';
+import { TextEditorValue } from '@/components/editor/model/types';
 
 export const useTextEditor = (initialValue = initialTextEditorValue) => {
   const [value, setValue] = useState(initialValue);
