+++ src/components/editor/styles/text-editor.css
